fix: misc fixes to rv32 execute e1

+++ extensions/rv32im/circuit/src/less_than/core.rs
@@ -340,7 +340,7 @@ where
 struct LessThanPreCompute {
     a: u8,
     b: u8,
-    c: u8,
+    c: u32,
 }
 
 impl<F, A, const LIMB_BITS: usize> StepExecutorE1<F>
@@ -380,7 +380,7 @@ where
         *pre_compute = LessThanPreCompute {
             a: a.as_canonical_u32() as u8,
             b: b.as_canonical_u32() as u8,
-            c: c.as_canonical_u32() as u8,
+            c: c.as_canonical_u32(),
         };
         let fn_ptr = match (e_u32 == RV32_IMM_AS, local_opcode == LessThanOpcode::SLTU) {
             (true, true) => execute_e1_impl::<_, _, true, true>,
@@ -405,9 +405,9 @@ unsafe fn execute_e1_impl<
 
     let rs1 = vm_state.vm_read::<u8, 4>(RV32_REGISTER_AS, pre_compute.b as u32);
     let rs2 = if E_IS_IMM {
-        imm_to_bytes(pre_compute.c as u32)
+        imm_to_bytes(pre_compute.c)
     } else {
-        vm_state.vm_read::<u8, 4>(RV32_REGISTER_AS, pre_compute.c as u32)
+        vm_state.vm_read::<u8, 4>(RV32_REGISTER_AS, pre_compute.c)
     };
     let cmp_result = if IS_U32 {
         u32::from_le_bytes(rs1) < u32::from_le_bytes(rs2)
